Support condition variables with keyword arguments
The motivation for this feature comes from a real world use case of this
library. We need to configure different rules per customer, and it makes
the most sense to leverage the Variables and Action objects that already exist.

Introducing kwargs to variable methods allows more customizability.
Defaults are set in code and configured values are used at runtime
if present. We opt for this approach over passing in parameters when
creating the Variables object because there are many Variables methods
which share common keyword args but must be configured independently.
Updating the business-rules library creates a cleaner interface than
managing dozens of verbose options during class instantiation.

This feature is backwards compatible with existing rules.